Kubernetes Client API. Retrieve number of nodes and cluster info

+++ platform-controller/src/main/java/org/hobbit/controller/kubernetes/K8sClusterManager.java
@@ -7,5 +7,5 @@ public interface K8sClusterManager {
 
     public V1PodList getPodsInfo() throws ApiException;
 
-    public int getNumberOfNodes();
+    public long getNumberOfNodes();
 }
